FE-65 Фоновое обновление счетчиков  подписок пользователей
